Disk encryption keys are now archived when host moves teams
